+++ src/Sharkable/SharkEndpoint/Metadata/SharkResponseMetadata.cs
@@ -1,0 +1,10 @@
+using Microsoft.AspNetCore.Http.Metadata;
+
+namespace Sharkable;
+
+internal sealed class SharkResponseMetadata : IProducesResponseTypeMetadata
+{
+    public int StatusCode { get; set; }
+    public Type? Type { get; set; }
+    public IEnumerable<string>? ContentTypes { get; set; }
+}